scripts/requirements.txt
langdetect==1.0.9
pdf2image==1.17.0
psycopg2==2.9.9
PyPDF2==3.0.1
pytesseract==0.3.10
python_magic==0.4.27
